bno055 orientation in rosbot

+++ ImuDriver.h
@@ -21,7 +21,7 @@ class ImuDriver :
 
         struct ImuMesurement 
         {
-            float orientation[4];
+            double orientation[4];
             float angular_velocity[3];
             float linear_velocity[3];
             uint32_t timestamp;
